Raspi/requirements.txt
pydbus
paho-mqtt